+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- storeConfigs query
+
 ## [2.90.8] - 2019-07-09
 ### Changed
 - Port OMSDataSource to a Janus Client.